==> 1.Data_Collection_from_API.ipynb <==
# SpaceX Falcon 9 Data Collection & Analysis

## 1. Data Retrieval and Processing
### Importing Necessary Libraries

import requests
import pandas as pd
import numpy as np
import datetime

pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', None)

### Define Auxilary Functions to Retrive Data

def getBoosterVersion(data):
    for x in data['rocket']:
       if x:
        response = requests.get("https://api.spacexdata.com/v4/rockets/"+str(x)).json()
        BoosterVersion.append(response['name'])

def getLaunchSite(data):
    for x in data['launchpad']:
       if x:
         response = requests.get("https://api.spacexdata.com/v4/launchpads/"+str(x)).json()
         Longitude.append(response['longitude'])
         Latitude.append(response['latitude'])
         LaunchSite.append(response['name'])

def getPayloadData(data):
    for load in data['payloads']:
       if load:
        response = requests.get("https://api.spacexdata.com/v4/payloads/"+load).json()
        PayloadMass.append(response['mass_kg'])
        Orbit.append(response['orbit'])

def getCoreData(data):
    for core in data['cores']:
            if core['core'] != None:
                response = requests.get("https://api.spacexdata.com/v4/cores/"+core['core']).json()
                Block.append(response['block'])
                ReusedCount.append(response['reuse_count'])
                Serial.append(response['serial'])
            else:
                Block.append(None)
                ReusedCount.append(None)
                Serial.append(None)
            Outcome.append(str(core['landing_success'])+' '+str(core['landing_type']))
            Flights.append(core['flight'])
            GridFins.append(core['gridfins'])
            Reused.append(core['reused'])
            Legs.append(core['legs'])
            LandingPad.append(core['landpad'])

### Retrieving SpaceX Data
#### Here if we get a 200 sucess code, that means data retrivel is sucess

spacex_url="https://api.spacexdata.com/v4/launches/past"
response = requests.get(spacex_url)
response.status_code

### DataFrame Setup

data = pd.json_normalize(response.json())
data.head()

## 2. Data Cleaning and Transformation
### Extracting Relevant Features
#### Subsetting the DataFrame to include only relevant features and filtering out rows with multiple cores and payloads.

# Lets take a subset of our dataframe keeping only the features we want and the flight number, and date_utc.
data = data[['rocket', 'payloads', 'launchpad', 'cores', 'flight_number', 'date_utc']]

# We will remove rows with multiple cores because those are falcon rockets with 2 extra rocket boosters and rows that have multiple payloads in a single rocket.
data = data[data['cores'].map(len)==1]
data = data[data['payloads'].map(len)==1]

# Since payloads and cores are lists of size 1 we will also extract the single value in the list and replace the feature.
data['cores'] = data['cores'].map(lambda x : x[0])
data['payloads'] = data['payloads'].map(lambda x : x[0])

# We also want to convert the date_utc to a datetime datatype and then extracting the date leaving the time
data['date'] = pd.to_datetime(data['date_utc']).dt.date

# Using the date we will restrict the dates of the launches
data = data[data['date'] <= datetime.date(2020, 11, 13)]

#### Create empty lists to store extracted data from raw dataframe

#Column Lists 
BoosterVersion = []
PayloadMass = []
Orbit = []
LaunchSite = []
Outcome = []
Flights = []
GridFins = []
Reused = []
Legs = []
LandingPad = []
Block = []
ReusedCount = []
Serial = []
Longitude = []
Latitude = []

### Data Extraction

getBoosterVersion(data)
getLaunchSite(data)
getPayloadData(data)
getCoreData(data)

BoosterVersion[0:5]

#### Create a dictionary to input extracted data

launch_dict = {'FlightNumber': list(data['flight_number']),
'Date': list(data['date']),
'BoosterVersion':BoosterVersion,
'PayloadMass':PayloadMass,
'Orbit':Orbit,
'LaunchSite':LaunchSite,
'Outcome':Outcome,
'Flights':Flights,
'GridFins':GridFins,
'Reused':Reused,
'Legs':Legs,
'LandingPad':LandingPad,
'Block':Block,
'ReusedCount':ReusedCount,
'Serial':Serial,
'Longitude': Longitude,
'Latitude': Latitude}

#### Create a dataframe using dictionary

# Create a data from launch_dict
data = pd.DataFrame.from_dict(launch_dict)
# Show the head of the dataframe
data.head()

#### Filter only Falcon 9 data

data_falcon9 = data[data['BoosterVersion']!='Falcon 1']
data_falcon9.loc[:,'FlightNumber'] = list(range(1, data_falcon9.shape[0]+1))
data_falcon9.head()

## 3. Data Wrangling 
### Dealing with Missing Values

data_falcon9.isnull().sum()

#### We have 2 columns with missing values

#### We handle Payload Mass missing values by assigning mean value

# Calculate the mean value of PayloadMass column
payloadmass_mean = data_falcon9['PayloadMass'].mean()
# Replace the np.nan values with its mean value
data_falcon9.loc[data_falcon9['PayloadMass'].isnull(), 'PayloadMass'] = payloadmass_mean
data_falcon9.isnull().sum()

## 4. Data Preprocessing
### Classify the Outcome of Missions

# landing_outcomes = values on Outcome column
landing_outcomes = data_falcon9["Outcome"].value_counts()
landing_outcomes

#### We need to classify these outcomes as binary value (Sucess, Failure)

for i,outcome in enumerate(landing_outcomes.keys()):
    print(i,outcome)

#### Classify the Failures

bad_outcomes=set(landing_outcomes.keys()[[1,3,5,6,7]])
bad_outcomes

#### Binary Labeling outcome

# landing_class = 0 if bad_outcome
# landing_class = 1 otherwise
landing_class = []
for x in data_falcon9["Outcome"]:
    if x in bad_outcomes:
        landing_class.append(0)
    else:
        landing_class.append(1)

landing_class[0:15]

data_falcon9['Class']=landing_class
data_falcon9.head(8)

data_falcon9.to_csv("Data_falcon9.csv", index=False)
