==> 2.Data_Collection_from_Web_Scrapping.ipynb <==
# Space X Data Web Scrapping and Analysis
## 1. Web Scrapping
### Importing Libraries

import sys

import requests
from bs4 import BeautifulSoup
import re
import unicodedata
import pandas as pd

### Functions for Data Extraction

def date_time(table_cells):
    """
    This function returns the data and time from the HTML  table cell
    Input: the  element of a table data cell extracts extra row
    """
    return [data_time.strip() for data_time in list(table_cells.strings)][0:2]

def booster_version(table_cells):
    """
    This function returns the booster version from the HTML  table cell 
    Input: the  element of a table data cell extracts extra row
    """
    out=''.join([booster_version for i,booster_version in enumerate( table_cells.strings) if i%2==0][0:-1])
    return out

def landing_status(table_cells):
    """
    This function returns the landing status from the HTML table cell 
    Input: the  element of a table data cell extracts extra row
    """
    out=[i for i in table_cells.strings][0]
    return out


def get_mass(table_cells):
    mass=unicodedata.normalize("NFKD", table_cells.text).strip()
    if mass:
        mass.find("kg")
        new_mass=mass[0:mass.find("kg")+2]
    else:
        new_mass=0
    return new_mass


def extract_column_from_header(row):
    """
    This function returns the landing status from the HTML table cell 
    Input: the  element of a table data cell extracts extra row
    """
    if (row.br):
        row.br.extract()
    if row.a:
        row.a.extract()
    if row.sup:
        row.sup.extract()
        
    colunm_name = ' '.join(row.contents)
    
    # Filter the digit and empty names
    if not(colunm_name.strip().isdigit()):
        colunm_name = colunm_name.strip()
        return colunm_name    

### Data Retrieval
#### If we get a 200 success code, data retrievel is success

response = requests.get("https://en.wikipedia.org/w/index.php?title=List_of_Falcon_9_and_Falcon_Heavy_launches&oldid=1027686922")
print(response.status_code)

#### Use BeautifulSoup to create objects

# Use BeautifulSoup() to create a BeautifulSoup object from a response text content
soup = BeautifulSoup(response.text,'html.parser')
soup.title

####  Extracting tables from the HTML content and identifying the target table for data extraction.

html_tables = soup.find_all('table')

first_launch_table = html_tables[2]
print(first_launch_table)

#### Extracting column names from the table headers using BeautifulSoup.

column_names = []

# Apply find_all() function with `th` element on first_launch_table
# Iterate each th element and apply the provided extract_column_from_header() to get a column name
# Append the Non-empty column name (`if name is not None and len(name) > 0`) into a list called column_names
find_th = first_launch_table.find_all('th')
for row in find_th:
    name = extract_column_from_header(row)
    if name is not None and len(name) > 0:
        column_names.append(name)

print(column_names)

### Data Extraction and Parsing
#### Parsing through table rows and cells to extract relevant data and store it in a dictionary.

launch_dict= dict.fromkeys(column_names)

# Remove an irrelvant column
del launch_dict['Date and time ( )']

# Let's initial the launch_dict with each value to be an empty list
launch_dict['Flight No.'] = []
launch_dict['Launch site'] = []
launch_dict['Payload'] = []
launch_dict['Payload mass'] = []
launch_dict['Orbit'] = []
launch_dict['Customer'] = []
launch_dict['Launch outcome'] = []
# Added some new columns
launch_dict['Version Booster']=[]
launch_dict['Booster landing']=[]
launch_dict['Date']=[]
launch_dict['Time']=[]

extracted_row = 0

# Extract each table 
for table_number, table in enumerate(soup.find_all('table', "wikitable plainrowheaders collapsible")):
    # get table row 
    for rows in table.find_all("tr"):
        # check to see if first table heading is as number corresponding to launch a number 
        if rows.th:
            if rows.th.string:
                flight_number = rows.th.string.strip()
                flag = flight_number.isdigit()
        else:
            flag = False
        # get table element 
        row = rows.find_all('td')
        # if it is number save cells in a dictionary 
        if flag:
            extracted_row += 1
            
            # Flight Number value
            launch_dict['Flight No.'].append(flight_number)
            
            # Date value
            datatimelist = date_time(row[0])
            date = datatimelist[0].strip(',')
            launch_dict['Date'].append(date)
            
            # Time value
            time = datatimelist[1]
            launch_dict['Time'].append(time)
              
            # Booster version
            bv = booster_version(row[1])
            if not bv:
                bv = row[1].a.string
            launch_dict['Version Booster'].append(bv)
            
            # Launch Site
            launch_site = row[2].a.string
            launch_dict['Launch site'].append(launch_site)
            
            # Payload
            payload = row[3].a.string
            launch_dict['Payload'].append(payload)
            
            # Payload Mass
            payload_mass = get_mass(row[4])
            launch_dict['Payload mass'].append(payload_mass)
            
            # Orbit
            orbit = row[5].a.string
            launch_dict['Orbit'].append(orbit)
            
            # Customer
            customer = row[6].a.string if row[6].a else None
            launch_dict['Customer'].append(customer)
            
            # Launch outcome
            launch_outcome = list(row[7].strings)[0]
            launch_dict['Launch outcome'].append(launch_outcome)
            
            # Booster landing
            booster_landing = landing_status(row[8])
            launch_dict['Booster landing'].append(booster_landing)

## 2.Data Cleaning and Preprocessing
### DataFrame Creation

df= pd.DataFrame({ key:pd.Series(value) for key, value in launch_dict.items() })
df

### Handling Missing Values
#### Checking for missing values in the DataFrame

df.isna().sum()

#### Since only small number (2) of rows contain missing values we drop it

df.dropna(inplace=True)

### Cleaning Data
#### Cleaning and standardizing values in specific columns

df["Launch outcome"].value_counts()

#### We have to deal with these duplicate success values

df["Launch outcome"] = df["Launch outcome"].replace("Success\n","Success")
df["Launch outcome"].value_counts()

df["Booster landing"].value_counts()

df.loc[df['Booster landing'] == "Failure "]

df["Booster landing"] = df["Booster landing"].replace("Failure ","Failure")
df["Booster landing"] = df["Booster landing"].replace("No attempt\n","No attempt")
df["Booster landing"].value_counts()

### Data Type Conversion
#### Converting data types of columns of 'Date', 'Time', and 'Payload mass' for consistency and analysis.

df['Date'] = pd.to_datetime(df['Date'])
df['Time'] = pd.to_datetime(df['Time'])
df['Time'] = df['Time'].dt.strftime('%H:%M:%S')
df.head()


pd.set_option('display.max_rows', None)
df["Payload mass"] = df["Payload mass"].astype(str)
df['Payload mass'].value_counts()

df['Payload mass'] = df['Payload mass'].str.replace(' kg', '')
df['Payload mass'] = df['Payload mass'].str.replace('~', '')
df['Payload mass'] = df['Payload mass'].str.replace(',', '')
df['Payload mass'] = df['Payload mass'].str.replace('5000–6000', '5500')
df.drop(df[df['Payload mass'] == "C"].index, inplace = True)
df.drop(df[df['Payload mass'] == "0"].index, inplace = True)
df['Payload mass'].value_counts()


df["Payload mass"] = df["Payload mass"].astype(int)
df.head()

df['Version Booster'].value_counts()

df["Version Booster"] = df["Version Booster"].replace("F9 B5 ♺","F9 B5")
df["Version Booster"] = df["Version Booster"].replace("F9 B5[268]","F9 B5")
df["Version Booster"] = df["Version Booster"].replace("F9 FT♺","F9 FT")
df['Version Booster'].value_counts()

df.drop(['Flight No.'],axis=1,inplace=True)
df.reset_index(inplace=True)
df.drop(['index'],axis=1,inplace=True)
df.head()

### Export the Data

df.to_csv("Data_SpaceX_WebScrapped.csv", index=False)
